pyproject.toml
[project]
name = "egile-mcp-notifier"
version = "0.1.0"
description = "MCP server for sending email notifications and managing Google Calendar events"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Egile Team" }
]
keywords = ["mcp", "notification", "email", "calendar", "google", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
    "uvicorn[standard]>=0.24.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.110.0",
]

[project.scripts]
egile-mcp-notifier = "egile_mcp_notifier.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/egile_mcp_notifier"]

[tool.black]
line-length = 100
